oversampling example

+++ hetero_linkneighbor_oversample.py
@@ -20,7 +20,7 @@ def run(rank, world_size, dataset):
     data = dataset[0]
 
     e1 = data.edge_types[0]
-    train_loader = (LinkNeighborLoader(data, edge_label_index=(e1, (data[e1].edge_index)), num_neighbors=[50, 50], batch_size=1024, num_workers=len(os.sched_getaffinity(0)) / 2))
+    train_loader = (LinkNeighborLoader(data, edge_label_index=(e1, (data[e1].edge_index)), num_neighbors=[50, 50], batch_size=1024, num_workers=int(len(os.sched_getaffinity(0)) / 2)))
 
 
     torch.manual_seed(12345)
